Provider better filesystem hinting
The hinting currently doesn't handle backup and restore phase
differently which can be improved.

This patch tells the kernel that the data won't be reused in the copy
function, thus it can discard it from the page cache after it was read.

Next, it will instead tell the kernel that the accused file and the tmp
file both will be needed for reading next. The first call to WILLNEED
should be put up further in the call chain to also support reading the
allocation map but this would invole tracking DONTNEED differently, so
it is not implemented here.

In the end, the page caches will be discarded explicitly by telling the
kernel that the data is no longer needed.

Testing shows that the active page cache during shaking now is released
after each file operation and system responsiveness improves during
shaking.