Fix calculation of last file ideal position
I think this is a bug so I fix this: It seems that this means to
calculate the ideal position of this file just before the last file in
the order. To do this correctly, we need to subtract y->size and not
y->blocks, as z->start is in bytes and not in blocks.