Add reflink support
Instead of doing the backup by copying, just try to reflink the accused
file. This makes a clone of the file data without actually copying the
data on filesystems that support it. If it fails, fall back to the old
behavior and don't try to reflink again.

We are not interested in fragmentation of the clone as it is going to be
deleted anyway, and even with fragmentation it is still good enough as
source for the rewrite phase. This saves 50% of IO to speed up the
process.